BiCubic Curve for E+ chiller part load modeling
This is a modified version of curve fit tool that comes with EnergyPlus
installation. Original tool can generate curve fits for quadratic, cubic
and biquadratic curve types. This tool has the additional capability to
generate bicubic curves. The primary goal for the development of this
tool was to create EIRfTPLR curves used for EnergyPlus reformulated EIR
chiller's part load modeling. However, this can be used for any other
application to create bicubic curve if there are two independent
variables and one dependent variable.